==> 1.kaeberlein04plos/step1.fit.kaeberlein04.Rmd <==
---
title: "Fit mutant with binomial aging model"
author: "h qin"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document: default
  pdf_document: default
---

```{r}
rm(list=ls())
library('flexsurv')
library('stringr')
source("../lifespan.r")
```

The RLS data used here are from Kaeberlein et al 2004, Plos Genetics 
http://journals.plos.org/plosbiology/article?id=10.1371/journal.pbio.0020296

The original data set is at
https://doi.org/10.1371/journal.pbio.0020296.sd002

```{r load_rls}
tb = read.csv("kaeberlein04_selected_strains.csv")
names(tb)
```

```{r init_report}
my.strains = c("BY4742", "sir2", "SIR2.ox", "fob1", "hxk2", "fob1 hxk2");
report = data.frame(my.strains)
report$samplesize = NA; report$R=NA; report$t0=NA; report$n=NA; report$G=NA; #report$longfilename=NA; 
#write.csv(tb, "kaeberlein04_selected_strains.csv", row.names = FALSE, quote = FALSE)
```

# Fit all RLS data sets by strains

```{r fitting, message=FALSE }
for( i in 1:length(report[,1])){
  my.data = tb[,i]
  my.data = my.data[! is.na(my.data)]
  report$samplesize[i] = length(my.data)
  
  GompFlex = flexsurvreg(formula = Surv(tb[,i]) ~ 1, dist = 'gompertz')
  WeibFlex = flexsurvreg(formula = Surv(tb[,i]) ~ 1, dist = 'weibull')

  report$avgLS[i] =  mean(tb[,i], na.rm=T)
  report$stdLS[i] =  sd(tb[,i], na.rm = T)
  report$CV[i] = report$stdLS[i] / report$avgLS[i]

  report$GompGFlex[i] = GompFlex$res[1,1]
  report$GompRFlex[i] = GompFlex$res[2,1]
  report$GompLogLikFlex[i] = round(GompFlex$loglik, 1)
  report$GompAICFlex[i] = round(GompFlex$AIC)

  report$WeibShapeFlex[i] = WeibFlex$res[1,1]
  report$WeibRateFlex[i] = WeibFlex$res[2,1]
  report$WeibLogLikFlex[i] = round(WeibFlex$loglik, 1)  
  report$WeibAICFlex[i] = round(WeibFlex$AIC)

  #set initial values
  Rhat = report$GompRFlex[i]; # 'i' was missing. a bug costed HQ a whole afternoon.
  Ghat = report$GompGFlex[i];
  nhat = 6;  
  t0= (nhat-1)/Ghat;
  fitBinom = optim ( c(Rhat, t0, nhat),  llh.binomialMortality.single.run, 
                     lifespan=tb[,i], 
                     #method='SANN') #SANN needs control  
                     method="L-BFGS-B", 
                     lower=c(1E-10, 1, 1), upper=c(1,200,20) );  
  report[i, c("R", "t0", "n")] = fitBinom$par[1:3]
  report$G[i] = (report$n[i] - 1)/report$t0[i]
}
```

Show the results
```{r report}
report
```


Output
```{r output}
write.csv(report, file = 'sandbox/_Kaeberlein04_singlefit.csv', row.names = FALSE)
```


# Overlay histogram with probability density function of  binomial aging model
see  http://hongqinlab.blogspot.com/2013/12/binomial-mortailty-model.html
 m = R ( 1 + t/t0)^(n-1)
 s = exp( (R t0/n)*(1 - (1+t/t0)^n ) )  
 pdf = s*m

It seems binomial model aging is a reasonable fit whenever Gompertz model is a reasonable fit. 
```{r histogram_overlay}

# html version, (x,y) are legend positions for plots
#x = c(0,    0,   0,     0,   40,   55,  50,   5);
#y = c(0.06, 0.06, 0.1, 0.1,0.04, 0.03, 0.03, 0.03  ); 

pdf("kae04_plots.pdf", width=4, height =4 );
# pdf version, (x,y) are legend positions for plots
#x = c(0,    32,   0,     0,   32,   0,  0,    0   );
#y = c(0.07, 0.06, 0.12, 0.12,0.05, 0.04, 0.04, 0.031  ); 


for( i in 1:length(report[,1])){
  my.data = tb[,i]
  my.data = my.data[! is.na(my.data)]
  
  h= hist(my.data, br=max(my.data)/2, plot = FALSE);
  hist(my.data, probability = TRUE, col='gray', border='white', xlab='RLS', ylab='probability',
       main=my.strains[i], ylim=c(0, max(h$density)*1.1), xlim=c(0, max(h$mids)*1.1))
  #plot( h$density ~ h$mids, main=my.strains[i], xlab="RLS",ylab="density")
  #par(new=TRUE);
  t= seq(0, max(h$mids)*1.2, by=0.1)
  s = exp( (report$R[i]* report$t0[i]/report$n[i])*(1 - (1+t/report$t0[i])^report$n[i] ) ) ; 
  m = report$R[i]*(1 + t/ report$t0[i])^(report$n[i] -1 )
  pdf = s*m
  lines( pdf ~ t, col='red', lty=1)
  
  s.g =  G.s( c(report$GompRFlex[i], report$GompGFlex[i]), t ); 
  m.g =  report$GompRFlex[i]*exp(report$GompGFlex[i]*t)
  pdf.g = s.g * m.g
  lines( pdf.g ~ t, col="blue", lty=2)
  
#  legend ( x[i], y[i], c("Binomial","Gompertz"), col=c("red","blue"), lty=c(1,2), cex=0.7 )
  legend ( "topright", c("Binomial","Gompertz"), col=c("red","blue"), lty=c(1,2), cex=0.7 )
}
dev.off()
```
