==> 1.kaeberlein04plos/step2.boostrap.kaeblein04.Rmd <==

---
title: "Bootstrap and fit mutants RLS with binomial aging model"
author: "H Qin"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document: default
  pdf_document: default
---

```{r}
rm(list=ls())
library('flexsurv')
library('stringr')
source("../lifespan.r")
set.seed(20170701) #for repeatability
RUNS = 100; #bootstrap runs

list.files()
```

# parse the strains from files

```{r}
tb = read.csv("kaeberlein04_selected_strains.csv")
summary(tb)
```

```{r}
strains = names(tb)
strains
```

```{r}
my.strains = c( "BY4742",  "sir2", "SIR2.ox",  
                "fob1","hxk2","fob1.hxk2"   )
names(tb) = my.strains
#write.csv(tb, file='kaeberlein04_selected_strains.csv', row.names = FALSE, quote=FALSE)

```

# Now, fit all RLS data 

```{r fit_all, message=FALSE }
for( BootstrapCount in 1:RUNS ) {#!!!!!!

report = data.frame(my.strains)
report$samplesize = NA; report$R=NA; report$t0=NA; report$n=NA; report$G=NA; #report$longfilename=NA; 

for( i in 1:length(report[,1])){
  my.data = tb[,i]
  my.data = my.data[! is.na(my.data)]
  report$samplesize[i] = length(my.data)
  
  #boostrap 
  bt.data = sample(my.data, replace=TRUE); # BOOTSTRAP HERE

  GompFlex = flexsurvreg(formula = Surv(bt.data) ~ 1, dist = 'gompertz')
  WeibFlex = flexsurvreg(formula = Surv(bt.data) ~ 1, dist = 'weibull')

  report$avgLS[i] =  mean(bt.data)
  report$stdLS[i] =  sd(bt.data)
  report$CV[i] = report$stdLS[i] / report$avgLS[i]

  report$GompGFlex[i] = GompFlex$res[1,1]
  report$GompRFlex[i] = GompFlex$res[2,1]
  report$GompLogLikFlex[i] = round(GompFlex$loglik, 1)
  report$GompAICFlex[i] = round(GompFlex$AIC)

  report$WeibShapeFlex[i] = WeibFlex$res[1,1]
  report$WeibRateFlex[i] = WeibFlex$res[2,1]
  report$WeibLogLikFlex[i] = round(WeibFlex$loglik, 1)  
  report$WeibAICFlex[i] = round(WeibFlex$AIC)

  #set initial values
  Rhat = report$GompRFlex[i]; # 'i' was missing. a bug costed HQ a whole afternoon.
  Ghat = report$GompGFlex[i];
  nhat = 6;  
  t0= (nhat-1)/Ghat;
  fitBinom = optim ( c(Rhat, t0, nhat),  llh.binomialMortality.single.run, 
                     lifespan=bt.data, 
                     #method='SANN') #SANN needs control  
                     method="L-BFGS-B", 
                     lower=c(1E-10, 1, 1), upper=c(1,200,20) );  
  report[i, c("R", "t0", "n")] = fitBinom$par[1:3]
  report$G[i] = (report$n[i] - 1)/report$t0[i]
  report$BinomialAIC[i] = 2*3 + 2*fitBinom$value #20171003 AIC for 3-parameter binomial aging model

}

report$BestModel = 'NA';
report$BestModel = ifelse((report$GompAICFlex < report$WeibAICFlex)&(report$GompAICFlex < report$BinomialAIC), "Gomp", report$BestModel)
report$BestModel = ifelse((report$BinomialAIC < report$WeibAICFlex)&(report$BinomialAIC < report$GompAICFlex), "Binom", report$BestModel)
report$BestModel = ifelse((report$WeibAICFlex < report$GompAICFlex)&(report$WeibAICFlex < report$BinomialAIC), "Weib", report$BestModel)
report$BestModel = ifelse( (abs(report$BinomialAIC - report$GompAICFlex)<2 ) | ( abs(report$BinomialAIC-report$WeibAICFlex) <2 ), "bin<2", report$BestModel)

outname = paste("bootstrap/", BootstrapCount, ".csv", sep='');  #has to be changed here !!!
write.csv(report, file = outname, row.names = FALSE);
}#end of boostrap loop

```

#summarize boostrap results. 
Pick row-col value from every file into a buffer, then mean and stddev. 

```{r summariz}
BootstrapMean = report; 
BootstrapStd = report;
rownames = names(report);
for( col in 2:length(report[1,])) {
  for ( row in 1:length(report[,1])) {
    buffer = c();
    for( BootstrapCount in 1:RUNS ) {
      filename = paste("bootstrap/", BootstrapCount, ".csv", sep='');
      tb = read.csv(filename)
      if( rownames[col] == "BestModel" ) {
        buffer = as.character( c(buffer, as.character(tb[row,col])) );
      } else {
        buffer = c(buffer, tb[row, col]); 
      }
    }
    if( rownames[col] == "BestModel" ) {
      tmp = table( buffer );
      BootstrapMean[row,col] = paste(names(tmp), tmp, sep="=", collapse = ":");
    } else {
     BootstrapMean[row,col] = mean(buffer);
     BootstrapStd[row,col] = sqrt(var(buffer));
    }
  }  
}
```

Means 
```{r}
BootstrapMean
```

StdDev
```{r}
BootstrapStd
```

Merge the two tables
```{r merging}
BootstrapMean$Rstd = BootstrapStd$R
BootstrapMean$t0std = BootstrapStd$t0
BootstrapMean$nstd = BootstrapStd$n
#BootstrapMean$Gstd = BootstrapStd$G
BootstrapMean$avgLSstd = BootstrapStd$avgLS
BootstrapMean$GompRFlexstd = BootstrapStd$GompRFlex
BootstrapMean$GompGFlexstd = BootstrapStd$GompGFlex
BootstrapMean$GompAICFlexstd = BootstrapStd$GompAICFlex
BootstrapMean$WeibAICFlexstd = BootstrapStd$WeibAICFlex
BootstrapMean$BinomialAICstd = BootstrapStd$BinomialAIC

BootstrapMean$avgLSstd = BootstrapStd$avgLS
names(BootstrapMean)
```


Reorganized the columns
```{r}
#BootstrapMean = BootstrapMean[, c( "my.strains",  "samplesize",  "R", "Rstd",  "t0",  "t0std",  "n",  #"nstd",   "G",  
#                                   "Gstd", "avgLS",  "avgLSstd", "stdLS",   "BestModel",  "CV",  
#                                    "GompRFlex", "GompRFlexStd", "GompGFlex",  "GompGFlexStd",
#                                   "GompLogLikFlex",   "WeibShapeFlex",  "WeibRateFlex", 
#                                   "WeibLogLikFlex",  "WeibAICFlex", "GompAICFlex", "BinomialAIC" 
#                                   )];

BootstrapMean = BootstrapMean[, c( "my.strains", "avgLS", "avgLSstd" , 
                                   "R", "Rstd",  "t0",  "t0std",  "n",  "nstd", "GompRFlex", "GompRFlexstd", "GompGFlex", "GompGFlexstd", "WeibAICFlex", "WeibAICFlexstd", "GompAICFlex",  "GompAICFlexstd", "BinomialAIC", "BinomialAICstd" 
 )];

BootstrapMean
```

Merge mean and std for publication
```{r}
BootstrapMeanPublishing = data.frame( BootstrapMean[, c("my.strains")] )
BootstrapMeanPublishing$avgLSwithStd = as.vector( paste(round(BootstrapMean$avgLS, 2), round(BootstrapMean$avgLSstd,3), sep=" +/- ") );
BootstrapMeanPublishing$RwithStd = as.vector( paste(round(BootstrapMean$R, 4), round(BootstrapMean$Rstd,4), sep=" +/- ") );
BootstrapMeanPublishing$t0withStd = as.vector( paste(round(BootstrapMean$t0, 1), round(BootstrapMean$t0std,1), sep=" +/- ") );
BootstrapMeanPublishing$nwithStd = as.vector( paste(round(BootstrapMean$n, 1), round(BootstrapMean$nstd,3), sep=" +/- ") );
#BootstrapMeanPublishing$GwithStd = as.vector( paste(round(BootstrapMean$G, 2), round(BootstrapMean$Gstd,3), sep=" +/- ") );

BootstrapMeanPublishing$GompR = as.vector( paste(round(BootstrapMean$GompRFlex, 4), round(BootstrapMean$GompRFlexstd,5), sep=" +/- ") );
BootstrapMeanPublishing$GompG = as.vector( paste(round(BootstrapMean$GompGFlex, 2), round(BootstrapMean$GompGFlexstd,3), sep=" +/- ") );
BootstrapMeanPublishing$WeibAIC = as.vector( paste(round(BootstrapMean$WeibAICFlex, 2), round(BootstrapMean$WeibAICFlexstd,2), sep=" +/- ") );
BootstrapMeanPublishing$GompAIC = as.vector( paste(round(BootstrapMean$GompAICFlex, 2), round(BootstrapMean$GompAICFlexstd,2), sep=" +/- ") );
BootstrapMeanPublishing$BinomialAIC = as.vector( paste(round(BootstrapMean$BinomialAIC, 2), round(BootstrapMean$BinomialAICstd,2), sep=" +/- ") );

BootstrapMeanPublishing
```

output
```{r output}
write.csv(BootstrapMean, file='sandbox/_Kaberlein04_Bootstrap_summary.csv', row.names = FALSE)
write.csv(BootstrapMeanPublishing, file='sandbox/_Kaberlein04_Bootstrap_summary_Publications.csv', row.names = FALSE)

```
